requirements.txt
streamlit
scikit-learn
pandas
numpy
plotly
joblib
Pillow

# AWS Integration
boto3
botocore
python-dotenv
